ActiveHealthCheckMonitor.Log nested type moved out to a separate file ActiveHealthCheckMonitor.Log.cs
